Remove social share and remove header (contributes to #1)
The social share buttons weren't that customizable. This meant that there were some h5's in there that didn't have any semantic meaning. Which is problematic in terms of accessibility.

And as I myself do not really care about those buttons, they were removed and the other H5 was replaced with a span element.